POST /register, POST /login 동작 확인 완

+++ Back/src/main/java/com/donggukReview/donggukReview/controller/MembershipController.java
@@ -2,12 +2,13 @@ package com.donggukReview.donggukReview.controller;
 
 import com.donggukReview.donggukReview.common.AuthUser;
 import com.donggukReview.donggukReview.dto.RegisterRequestDto;
+import com.donggukReview.donggukReview.dto.WithdrawRequestDTO;
 import com.donggukReview.donggukReview.entity.Users;
 import com.donggukReview.donggukReview.service.UserService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import lombok.extern.slf4j.Slf4j;
@@ -21,16 +22,25 @@ public class MembershipController {
 	@Autowired
 	private UserService userService;
 
+	@Autowired
+	private BCryptPasswordEncoder encoder;
+
 	@PostMapping("/register")
 	public ResponseEntity<?> register(
-			@RequestPart(value = "data", required = true) RegisterRequestDto requestDto,
+			@RequestPart(value = "data") RegisterRequestDto requestDto,
 			@RequestPart(value = "file", required = false) MultipartFile file
 	) {
 		if (userService.existsByUserId(requestDto.getUserId())) {
 			return ResponseEntity.badRequest().body("이미 존재하는 회원입니다.");
 		}
 
-		long createdUserId = userService.register(requestDto, file);
+		Users userEntity = new Users();
+		userEntity.setUserId(requestDto.getUserId());
+		userEntity.setUserNickname(requestDto.getUserNickname());
+		userEntity.setUserPassword(encoder.encode(requestDto.getUserPassword()));
+		userEntity.setUserRole("ROLE_USER");
+
+		long createdUserId = userService.register(userEntity, file);
 
 		if (createdUserId == -1) {
 			return ResponseEntity.internalServerError().body("회원 등록에 실패하였습니다.");
@@ -40,9 +50,13 @@ public class MembershipController {
 	}
 
 	@DeleteMapping("/withdraw")
-	public ResponseEntity<?> register(@AuthUser Users userEntity) {
+	public ResponseEntity<?> register(@AuthUser Users userEntity, @RequestBody WithdrawRequestDTO requestDto) {
 		if (userEntity == null) {
 			return ResponseEntity.notFound().build();
+		}
+
+		if (!requestDto.checkPassword() || !encoder.matches(requestDto.getUserPassword(), userEntity.getUserPassword())) {
+			return ResponseEntity.badRequest().body("비밀번호를 다시 확인하세요.");
 		}
 
 		userService.deleteUser(userEntity.getId());
